backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==41.0.7

# Authentication
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==3.2.2

# Document parsing
python-docx==1.1.0

# Testing (optional, for development)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0

aiofiles
# wechat
wechatpayv3
openai